[DEV-6264] Split events_controller test module
